# modules/mongoModule/models/iperf_result_model_mongo.py
class IperfResultModelMongo:
    def __init__(self, 
                 msm_id, repetition_number, start_timestamp,
                 transport_protocol, source_ip, source_port, destination_ip,
                 destination_port, bytes_received, duration, avg_speed, full_result,
                 type, timestamp_iso):
        self._id = None
        self.msm_id = msm_id
        self.repetition_number = repetition_number
        self.transport_protocol = transport_protocol
        self.start_timestamp = start_timestamp
        self.source_ip = source_ip
        self.source_port = source_port
        self.destination_ip = destination_ip
        self.destination_port = destination_port
        self.bytes_received = bytes_received
        self.duration = duration
        self.avg_speed = avg_speed
        self.full_result = full_result
        self.type = type
        self.timestamp_iso = timestamp_iso


    def to_dict(self):
        return {
            'msm_id': self.msm_id,
            'repetition_number': self.repetition_number,
            'transport_protocol': self.transport_protocol,
            'start_timestamp': self.start_timestamp,
            'bytes_received': self.bytes_received,
            'source_port': self.source_port,
            'destination_port': self.destination_port,
            'duration': self.duration,
            'avg_speed': self.avg_speed,
            'full_result': self.full_result,
            'type': self.type,
            'timestamp_iso': self.timestamp_iso
        }